Trying to consolidate common code rather than extract into a separate class.
